+++ priv/repo/migrations/20190220221350_add_pending_location_table.exs
@@ -6,8 +6,8 @@ defmodule BlockchainAPI.Repo.Migrations.AddPendingLocationTable do
       add :hash, :string, null: false
       add :status, :string, null: false, default: "pending"
       add :location, :string, null: false
-      add :fee, :integer, null: false
-      add :nonce, :integer, null: false
+      add :fee, :bigint, null: false
+      add :nonce, :bigint, null: false
 
       add :owner, references(:accounts, on_delete: :nothing, column: :address, type: :string), null: false
       add :gateway, references(:gateway_transactions, on_delete: :nothing, column: :gateway, type: :string), null: false
